update and fix
