+++ nishiki_koi_shop/src/assets/css/Home.css
@@ -53,7 +53,7 @@
     background-color: #f9f9f9;
 }
 
-.formAdContent {
+.farmAdContent {
     display: flex;
     justify-content: space-between;
     align-items: center;
