+++ nishiki_koi_shop/src/pages/Tour/Tour.jsx
@@ -1,5 +1,4 @@
-import {useParams} from "react-router-dom";
-import {useContext, useEffect, useState} from "react";
+import {useState} from "react";
 import '../../assets/css/Tour/tour.css'
 
 const Tour = () => {
