+++ nishiki_koi_shop/src/store/sampleTest.jsx
@@ -56,7 +56,7 @@ export const tempTour = [
     },
 ]
 
-export const tempFarm=[
+export const tempFarm = [
     {
         id: 'KOI-E-3306-11-01',
         name: 'Natalya Farm',
@@ -77,3 +77,109 @@ export const tempFarm=[
         maxJoiner: 60,
     },
 ]
+
+export const koiProducts = [
+    {
+        id: 'KF-2203-3303-1',
+        name: 'Koi Fish 1',
+        description: 'Beautiful Koi 1',
+        price: '$200',
+        imageUrl: 'https://media.la34.com.vn/upload/image/201810/medium/46263_koi2.jpg'
+    },
+    {
+        id: 'KF-2203-3303-2',
+        name: 'Koi Fish 2',
+        description: 'Beautiful Koi 2',
+        price: '$250',
+        imageUrl: 'https://bizweb.dktcdn.net/100/004/358/files/53-copy.jpg?v=1480156373075'
+    },
+    {
+        id: 'KF-2203-3303-3',
+        name: 'Koi Fish 3',
+        description: 'Beautiful Koi 3',
+        price: '$300',
+        imageUrl: 'https://bizweb.dktcdn.net/100/004/358/files/75-copy.jpg?v=1480156644996'
+    },
+    {
+        id: 'KF-2203-3303-4',
+        name: 'Koi Fish 4',
+        description: 'Beautiful Koi 4',
+        price: '$350',
+        imageUrl: 'https://bizweb.dktcdn.net/100/004/358/files/72-copy.jpg?v=1480156608949'
+    },
+    {
+        id: 'KF-2203-3303-5',
+        name: 'Koi Fish 5',
+        description: 'Beautiful Koi 5',
+        price: '$400',
+        imageUrl: 'https://media.la34.com.vn/upload/image/201810/medium/46263_koi2.jpg'
+    },
+    {
+        id: 'KF-2203-3303-6',
+        name: 'Koi Fish 6',
+        description: 'Beautiful Koi 6',
+        price: '$450',
+        imageUrl: 'https://bizweb.dktcdn.net/100/004/358/files/53-copy.jpg?v=1480156373075'
+    },
+    {
+        id: 'KF-2203-3303-7',
+        name: 'Koi Fish 7',
+        description: 'Beautiful Koi 7',
+        price: '$500',
+        imageUrl: 'https://bizweb.dktcdn.net/100/004/358/files/75-copy.jpg?v=1480156644996'
+    },
+    {
+        id: 'KF-2203-3303-8',
+        name: 'Koi Fish 8',
+        description: 'Beautiful Koi 8',
+        price: '$550',
+        imageUrl: 'https://bizweb.dktcdn.net/100/004/358/files/72-copy.jpg?v=1480156608949'
+    },
+    {
+        id: 'KF-2203-3303-9',
+        name: 'Koi Fish 9',
+        description: 'Beautiful Koi 9',
+        price: '$600',
+        imageUrl: 'https://media.la34.com.vn/upload/image/201810/medium/46263_koi2.jpg'
+    },
+    {
+        id: 'KF-2203-3303-10',
+        name: 'Koi Fish 10',
+        description: 'Beautiful Koi 10',
+        price: '$650',
+        imageUrl: 'https://bizweb.dktcdn.net/100/004/358/files/53-copy.jpg?v=1480156373075'
+    },
+    {
+        id: 'KF-2203-3303-11',
+        name: 'Koi Fish 11',
+        description: 'Beautiful Koi 11',
+        price: '$700',
+        imageUrl: 'https://bizweb.dktcdn.net/100/004/358/files/75-copy.jpg?v=1480156644996'
+    },
+    {
+        id: 'KF-2203-3303-12',
+        name: 'Koi Fish 12',
+        description: 'Beautiful Koi 12',
+        price: '$750',
+        imageUrl: 'https://bizweb.dktcdn.net/100/004/358/files/72-copy.jpg?v=1480156608949'
+    },
+];
+
+export const tripItems = [
+    {
+        farmName: 'Nông trại A',
+        imageUrl: 'https://i.ytimg.com/vi/pUADEpL3DNM/maxresdefault.jpg',
+    },
+    {
+        farmName: 'Nông trại B',
+        imageUrl: 'https://www.kodamakoifarm.com/wp-content/uploads/2021/12/Taro.jpg',
+    },
+    {
+        farmName: 'Nông trại C',
+        imageUrl: 'https://www.kodamakoifarm.com/wp-content/uploads/2019/07/taro-koi-japan-v1.jpg',
+    },
+    {
+        farmName: 'Nông trại D',
+        imageUrl: 'https://www.koikoimatsuda.jp/en/set_pics/home02.JPG',
+    },
+];
